+++ mod-source-record-manager-server/src/main/java/org/folio/services/afterprocessing/InstanceProcessingServiceImpl.java
@@ -1,19 +1,26 @@
 package org.folio.services.afterprocessing;
 
-import io.vertx.core.CompositeFuture;
+import javax.ws.rs.NotFoundException;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
+
 import io.vertx.core.Future;
 import io.vertx.core.http.HttpMethod;
 import io.vertx.core.json.JsonObject;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
 import org.apache.commons.collections4.CollectionUtils;
-import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang3.tuple.Pair;
 import org.folio.HttpStatus;
 import org.folio.dao.JobExecutionSourceChunkDao;
 import org.folio.dataimport.util.OkapiConnectionParams;
 import org.folio.dataimport.util.RestUtil;
 import org.folio.rest.client.SourceStorageClient;
 import org.folio.rest.jaxrs.model.Instance;
+import org.folio.rest.jaxrs.model.Instances;
 import org.folio.rest.jaxrs.model.JobExecutionSourceChunk;
 import org.folio.rest.jaxrs.model.ParsedRecord;
 import org.folio.rest.jaxrs.model.ParsedRecordCollection;
@@ -24,16 +31,11 @@ import org.folio.services.parsers.RecordFormat;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import javax.ws.rs.NotFoundException;
-import java.util.ArrayList;
-import java.util.List;
-
 @Service
 public class InstanceProcessingServiceImpl implements AfterProcessingService {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(InstanceProcessingServiceImpl.class);
-  private static final String INVENTORY_URL = "/inventory/instances";
-  private static final String INSTANCE_LOCATION_RESPONSE_HEADER = "location";
+  private static final String INVENTORY_URL = "/inventory/instances/batch";
 
   private JobExecutionSourceChunkDao jobExecutionSourceChunkDao;
   private AdditionalFieldsUtil additionalInstanceFieldsUtil;
@@ -46,27 +48,14 @@ public class InstanceProcessingServiceImpl implements AfterProcessingService {
 
   @Override
   public Future<Void> process(List<Record> records, String sourceChunkId, OkapiConnectionParams params) {
-    if (CollectionUtils.isEmpty(records)) {
-      return Future.succeededFuture();
-    }
     Future<Void> future = Future.future();
-    List<Future> processRecordFutures = new ArrayList<>();
-    RecordToInstanceMapper mapper = RecordToInstanceMapperBuilder.buildMapper(RecordFormat.getByDataType(getRecordsType(records)));
-    records
-      .parallelStream()
-      .forEach(record -> processRecordFutures.add(processRecordForInstance(record, mapper, params)));
-    CompositeFuture.all(processRecordFutures).setHandler(result -> {
-      if (result.failed()) {
-        LOGGER.error("Couldn't create Instance in mod-inventory", result.cause());
-        jobExecutionSourceChunkDao.getById(sourceChunkId, params.getTenantId())
-          .compose(optional -> optional
-            .map(sourceChunk -> jobExecutionSourceChunkDao.update(sourceChunk.withState(JobExecutionSourceChunk.State.ERROR), params.getTenantId()))
-            .orElseThrow(() -> new NotFoundException(String.format(
-              "Couldn't update failed jobExecutionSourceChunk status to ERROR, jobExecutionSourceChunk with id %s was not found", sourceChunkId))));
+    List<Pair<Record, Instance>> recordToInstanceList = mapRecordsToInstances(records);
+    List<Instance> instances = recordToInstanceList.parallelStream().map(Pair::getValue).collect(Collectors.toList());
+    postInstances(instances, params).setHandler(ar -> {
+      if (ar.failed()) {
+        updateSourceChunkState(sourceChunkId, JobExecutionSourceChunk.State.ERROR, params);
       } else {
-        if (Record.RecordType.MARC.equals(records.get(0).getRecordType())) {
-          updateParsedRecords(records, params);
-        }
+        addAdditionalFields(recordToInstanceList, params);
       }
       // Complete future in order to continue the import process regardless of the result of creating Instances
       future.complete();
@@ -75,66 +64,109 @@ public class InstanceProcessingServiceImpl implements AfterProcessingService {
   }
 
   /**
-   * Performs mapping a given Record to Instance and sends result Instance to mod-inventory.
-   * If given Record is MARC record, adds additional fields into ParsedRecord.
-   * If given Record is ERROR record - just successfully complete.
+   * Performs mapping a given Records to Instances.
    *
-   * @param record              Record
-   * @param mapper              Instance mapper
-   * @param params              OKAPI connection paras
-   * @return future
+   * @param records given list of records
+   * @return association between Records and corresponding Instances
    */
-  private Future<Void> processRecordForInstance(Record record, RecordToInstanceMapper mapper, OkapiConnectionParams params) {
-    // If given Record is not ERROR Record - try to map Record to Instance
-    if (record.getParsedRecord() != null && record.getParsedRecord().getContent() != null) {
-      try {
-        Instance instance = mapper.mapRecord(new JsonObject(record.getParsedRecord().getContent().toString()));
-        return postInstance(instance, params).compose(instanceId -> {
-          if (Record.RecordType.MARC.equals(record.getRecordType())) {
-            additionalInstanceFieldsUtil.addInstanceIdToMarcRecord(record, instanceId);
-          }
-          return Future.succeededFuture();
-        });
-      } catch (Exception exception) {
-        String errorMessage =
-          String.format("Can not process given Record for Instance. Cause: '%s'. Exception: '%s'", exception.getMessage(), exception);
-        LOGGER.error(errorMessage);
-        return Future.failedFuture(errorMessage);
-      }
+  private List<Pair<Record, Instance>> mapRecordsToInstances(List<Record> records) {
+    if (CollectionUtils.isEmpty(records)) {
+      return Collections.emptyList();
     }
-    // IF given Record is ERROR record - return succeeded future
-    return Future.succeededFuture();
+    final RecordToInstanceMapper mapper = RecordToInstanceMapperBuilder.buildMapper(RecordFormat.getByDataType(getRecordsType(records)));
+    return records.parallelStream()
+      .map(record -> mapRecordToInstance(mapper, record))
+      .filter(Objects::nonNull)
+      .collect(Collectors.toList());
   }
 
   /**
-   * Creates Inventory Instance
+   * Maps a record to an instance record.
    *
-   * @param instance Instance entity serves as request body to POST request to Inventory
-   * @param params   params enough to connect ot OKAPI
-   * @return future with Instance id
+   * @param mapper a record to instance record mapper.
+   * @param record a record.
+   * @return either a pair of record-instance or null.
    */
-  private Future<String> postInstance(Instance instance, OkapiConnectionParams params) {
-    Future<String> future = Future.future();
-    RestUtil.doRequest(params, INVENTORY_URL, HttpMethod.POST, instance).setHandler(responseResult -> {
+  private Pair<Record, Instance> mapRecordToInstance(RecordToInstanceMapper mapper, Record record) {
+    try {
+      if (record.getParsedRecord() != null && record.getParsedRecord().getContent() != null) {
+        Instance instance = mapper.mapRecord(new JsonObject(record.getParsedRecord().getContent().toString()));
+        return Pair.of(record, instance);
+      }
+    } catch (Exception exception) {
+      String errorMessage = String.format("Can not map a given Record to Instance. Cause: '%s'. Exception: '%s'", exception.getMessage(), exception);
+      LOGGER.error(errorMessage);
+    }
+    return null;
+  }
+
+  /**
+   * Sends given collection of Instances to mod-inventory
+   *
+   * @param instanceList collection of Instances
+   * @param params       Okapi connection params
+   * @return future
+   */
+  private Future<Void> postInstances(List<Instance> instanceList, OkapiConnectionParams params) {
+    if (CollectionUtils.isEmpty(instanceList)) {
+      return Future.succeededFuture();
+    }
+    Future<Void> future = Future.future();
+    Instances instances = new Instances().withInstances(instanceList).withTotalRecords(instanceList.size());
+    RestUtil.doRequest(params, INVENTORY_URL, HttpMethod.POST, instances).setHandler(responseResult -> {
       try {
         if (!RestUtil.validateAsyncResult(responseResult, future)) {
-          LOGGER.error("Error creating new Instance record", future.cause());
+          LOGGER.error("Error creating a new collection of Instances", future.cause());
         } else {
-          if (StringUtils.isNotEmpty(instance.getId())) {
-            String instanceId = instance.getId();
-            future.complete(instanceId);
-          } else {
-            String location = responseResult.result().getResponse().getHeader(INSTANCE_LOCATION_RESPONSE_HEADER);
-            String instanceId = location.substring(location.lastIndexOf('/') + 1);
-            future.complete(instanceId);
-          }
+          future.complete();
         }
       } catch (Exception e) {
-        LOGGER.error("Error during post for new Instance", e);
+        LOGGER.error("Error during post for new collection of Instances", e);
         future.fail(e);
       }
     });
     return future;
+  }
+
+  /**
+   * Adds additional custom fields to parsed records and updates parsed records in mod-source-record-storage
+   *
+   * @param recordToInstanceList association between Records and corresponding Instances
+   * @param params               okapi connection params
+   */
+  private void addAdditionalFields(List<Pair<Record, Instance>> recordToInstanceList, OkapiConnectionParams params) {
+    if (CollectionUtils.isEmpty(recordToInstanceList)) {
+      return;
+    }
+
+    if (Record.RecordType.MARC == recordToInstanceList.get(0).getKey().getRecordType()) {
+      List<Record> records = recordToInstanceList.parallelStream()
+        .peek(it -> additionalInstanceFieldsUtil.addInstanceIdToMarcRecord(it.getKey(), it.getValue().getId()))
+        .map(Pair::getKey)
+        .collect(Collectors.toList());
+
+      updateParsedRecords(records, params).setHandler(updatedAr -> {
+        if (updatedAr.failed()) {
+          LOGGER.error("Couldn't update parsed records", updatedAr.cause());
+        }
+      });
+    }
+  }
+
+  /**
+   * Updates state of given source chunk
+   *
+   * @param sourceChunkId id of source chunk
+   * @param state         state of source chunk
+   * @param params        okapi connection params
+   */
+  private void updateSourceChunkState(String sourceChunkId, JobExecutionSourceChunk.State state, OkapiConnectionParams params) {
+    jobExecutionSourceChunkDao.getById(sourceChunkId, params.getTenantId())
+      .compose(optional -> optional
+        .map(sourceChunk -> jobExecutionSourceChunkDao.update(sourceChunk.withState(state), params.getTenantId()))
+        .orElseThrow(() ->
+          new NotFoundException(
+            String.format("Couldn't update failed jobExecutionSourceChunk status to ERROR, jobExecutionSourceChunk with id %s was not found", sourceChunkId))));
   }
 
   /**
@@ -151,8 +183,8 @@ public class InstanceProcessingServiceImpl implements AfterProcessingService {
    * Updates a collection of parsedRecords
    *
    * @param records collection of records containing updated parsed records
-   * @param params OkapiConnectionParams
-   * @return void
+   * @param params  okapi connection params
+   * @return Future
    */
   private Future<Void> updateParsedRecords(List<Record> records, OkapiConnectionParams params) {
     if (CollectionUtils.isEmpty(records)) {
@@ -160,29 +192,34 @@ public class InstanceProcessingServiceImpl implements AfterProcessingService {
     }
     Future<Void> future = Future.future();
     try {
-      List<ParsedRecord> parsedRecords = new ArrayList<>();
-      records.forEach(record -> {
-        if (record.getParsedRecord() != null) {
-          parsedRecords.add(record.getParsedRecord());
-      }});
-      ParsedRecordCollection parsedRecordsCollection = new ParsedRecordCollection()
-        .withParsedRecords(parsedRecords)
-        .withTotalRecords(parsedRecords.size())
-        .withRecordType(ParsedRecordCollection.RecordType.valueOf(records.get(0).getRecordType().value()));
-      SourceStorageClient sourceStorageClient = new SourceStorageClient(params.getOkapiUrl(), params.getTenantId(), params.getToken());
-      sourceStorageClient.putSourceStorageParsedRecordsCollection(parsedRecordsCollection, response -> {
-        if (response.statusCode() != HttpStatus.HTTP_OK.toInt()) {
-          String errorMessage = String.format("Couldn't update parsed records collection - response status code %s, expected 200", response.statusCode());
-          LOGGER.error(errorMessage);
-          future.fail(errorMessage);
-        } else {
-          future.complete();
-        }
-      });
+      new SourceStorageClient(params.getOkapiUrl(), params.getTenantId(), params.getToken())
+        .putSourceStorageParsedRecordsCollection(buildParsedRecordCollection(records), response -> {
+          if (HttpStatus.HTTP_OK.toInt() != response.statusCode()) {
+            setFail(future, response.statusCode());
+          }
+        });
     } catch (Exception e) {
       LOGGER.error("Failed to update parsed records collection", e);
       future.fail(e);
     }
-    return future;
+    return future.isComplete() ? future : Future.succeededFuture();
+  }
+
+  private void setFail(Future<Void> future, int statusCode) {
+    String errorMessage = String.format("Couldn't update parsed records collection - response status code %s, expected 200", statusCode);
+    LOGGER.error(errorMessage);
+    future.fail(errorMessage);
+  }
+
+  private ParsedRecordCollection buildParsedRecordCollection(List<Record> records) {
+    List<ParsedRecord> parsedRecords = records.stream()
+      .filter(record -> record.getParsedRecord() != null)
+      .map(Record::getParsedRecord)
+      .collect(Collectors.toList());
+
+    return new ParsedRecordCollection()
+      .withParsedRecords(parsedRecords)
+      .withTotalRecords(parsedRecords.size())
+      .withRecordType(ParsedRecordCollection.RecordType.valueOf(getRecordsType(records).value()));
   }
 }